feat: extract kg build cli into fancyrag package
