+++ docs/prd.md
@@ -5,6 +5,7 @@
 - Deliver a self-contained GraphRAG evaluation stack that runs on a single Linux host using Docker Compose for Neo4j and Qdrant plus the official `neo4j-graphrag` package.
 - Provide scripted workflows that create the Neo4j vector index, build the knowledge graph, mirror embeddings into Qdrant, and execute retrieval checks end-to-end.
 - Maintain compatibility with production-bound settings so the same scripts can later point at managed services without code changes.
+- Improve maintainability by decomposing the `scripts/kg_build.py` monolith into a modular `src/fancyrag/` package with clear testing guardrails.
 
 ### Background Context
 Earlier iterations assumed existing managed Neo4j and Qdrant deployments. That scope stalled adoption because teams could not safely experiment without access to those services. Version 1 focuses on a project-owned stack: Dockerized Neo4j 5.26.12 (APOC Core) and Qdrant 1.15.4 (see [Version Matrix](../README.md#version-matrix)), paired with Python 3.12 tooling that relies exclusively on the public `neo4j-graphrag` APIs (`SimpleKGPipeline`, `create_vector_index`, `QdrantNeo4jRetriever`). Once the local workflow is proven, the same scripts will target managed infrastructure by overriding connection variables.
@@ -14,6 +15,7 @@ Earlier iterations assumed existing managed Neo4j and Qdrant deployments. That s
 |------------|---------|----------------------------------------------------------------|------------|
 | 2025-09-24 | 0.1     | Rebuilt PRD in BMAD format; aligned goals, requirements, epics | Codex CLI  |
 | 2025-09-28 | 0.2     | Narrowed scope to local Docker stack + scripted minimal path   | Codex CLI  |
+| 2025-10-02 | 0.3     | Added refactor roadmap and linked planning artifacts          | Codex CLI  |
 
 ## Requirements
 ### Functional
@@ -52,6 +54,8 @@ Earlier iterations assumed existing managed Neo4j and Qdrant deployments. That s
 |------|-------------------------------------|----------------------------------------------------------------------------------------|
 | 1    | Environment & Workspace             | Establish reproducible Python environment and configuration surface for local stack.   |
 | 2    | Local GraphRAG Minimal Path         | Deliver Dockerized Neo4j/Qdrant plus scripts covering KG build, vector sync, retrieval. |
+| 3    | Ingestion Quality Hardening         | Add adaptive chunking, QA telemetry, and semantic enrichment for reliable ingestion.    |
+| 4    | FancyRAG `kg_build.py` Refactor     | Break the KG build monolith into modular packages with dedicated tests and guardrails. |
 
 ## Epic Details
 ### Epic 1 — Environment & Workspace
@@ -65,3 +69,20 @@ Earlier iterations assumed existing managed Neo4j and Qdrant deployments. That s
 - Add `docker-compose.neo4j-qdrant.yml` with persistent volumes and documented environment overrides.
 - Provide a script that runs `create_vector_index` and a `SimpleKGPipeline` ingestion with sample content.
 - Provide export and retrieval scripts that push embeddings to Qdrant and execute `GraphRAG.search()` to verify grounded answers.
+
+
+### Epic 3 — Ingestion Quality Hardening
+**Goal:** Raise ingestion fidelity with adaptive chunking, QA gating, and semantic enrichment so telemetry can block bad data before retrieval.
+- Introduce chunking presets and deterministic directory ingestion with metadata captured for observability.
+- Expand QA evaluation to emit JSON/Markdown reports and enforce thresholds tied to CI checks.
+- Wire semantic enrichment toggles that leverage GraphRAG `LLMEntityRelationExtractor` while maintaining rollback controls.
+- Status: Delivered 2025-10-02 (see `docs/bmad/focused-epics/ingestion-quality/epic.md`).
+
+### Epic 4 — FancyRAG `kg_build.py` Monolith Decomposition
+**Goal:** Break the `scripts/kg_build.py` script into a `src/fancyrag/` package with dedicated modules for CLI wiring, pipeline orchestration, QA utilities, Neo4j helpers, and configuration.
+- Follow the guardrails defined in the project brief and architecture addendum (single responsibility modules, typed interfaces, 200–400 LOC targets).
+- Keep `scripts/kg_build.py` as a thin CLI bridge that defers to `src/fancyrag/cli/kg_build_main.py`.
+- Stand up per-module unit tests plus a CLI smoke test that proves parity with the pre-refactor behaviour.
+- Planning artifacts: `docs/prd/projects/fancyrag-kg-build-refactor/`, `docs/architecture/projects/fancyrag-kg-build-refactor.md`, and `docs/bmad/focused-epics/kg-build-refactor/epic.md`.
+- Status: Planned (kick-off after Epic 3 hardening).
+
