+++ docs/prd/epics.md
@@ -5,3 +5,4 @@
 | 1    | Environment & Workspace     | Establish reproducible Python environment and configuration surface for local stack.   |
 | 2    | Local GraphRAG Minimal Path | Deliver Dockerized Neo4j/Qdrant plus scripts covering KG build, vector sync, retrieval. |
 | 3    | Ingestion Quality Hardening | Add adaptive chunking, ingestion QA, and semantic enrichment for reliable GraphRAG pipelines. |
+| 4    | FancyRAG `kg_build.py` Refactor | Break the KG build monolith into modular packages with CLI-smoke test coverage. |
