+++ docs/qa/assessments/4.1-nfr-20251003.md
@@ -1,0 +1,40 @@
+# NFR Assessment – Story 4.1 (CLI Bridge for FancyRAG kg_build)
+
+Date: 2025-10-03
+Assessor: Quinn (QA)
+
+## Security
+- **Status:** PASS
+- `ensure_env` is invoked before establishing clients, preventing silent execution when env secrets are absent (`src/fancyrag/cli/kg_build_main.py:2099-2107`).
+- Sanitization continues to rely on `cli.sanitizer.scrub_object`, so run logs do not leak credentials.
+- No new environment variables or persisted secrets were introduced.
+
+## Performance
+- **Status:** PASS
+- The CLI still executes the same pipeline path; argument defaults and batching logic remain unchanged.
+- Optional semantic enrichment limits concurrency via existing flag plumbing, so no regression expected.
+- `scripts/kg_build.py --help` parity snapshots confirm no new heavyweight defaults were added.
+
+## Reliability
+- **Status:** PASS
+- Wrapper delegates to the packaged entry point, eliminating duplicate logic and reducing drift (`scripts/kg_build.py:1-9`).
+- Unit tests mock the pipeline to ensure env checks happen before execution and that the wrapper is correctly wired (`tests/unit/fancyrag/cli/test_kg_build_main.py`).
+- Existing kg_build test suite still passes, exercising rollback and QA threshold behaviour (`tests/unit/scripts/test_kg_build.py`).
+
+## Maintainability
+- **Status:** PASS
+- Source tree documentation now records the FancyRAG CLI package, aiding discoverability (`docs/architecture/source-tree.md:23-74`).
+- The change aligns the codebase with the planned FancyRAG module layout, clearing the path for the remaining epic stories.
+- Story Dev Notes and completion log capture artefact locations and command history for future reference.
+
+## Observability
+- **Status:** PASS
+- Log formatting and QA reporting remain untouched, and parity snapshots show operator-facing help text unchanged.
+- New unit tests assert the presence of env validation, preventing silent misconfiguration scenarios that would otherwise surface only at runtime.
+
+## Recommendations
+- Integrate the CLI help snapshot step into CI to guard against future flag regressions (covers risk TECH-041).
+- When Story 4.2 extracts the pipeline module, ensure its unit tests continue to import through `fancyrag.cli.kg_build_main` to maintain end-to-end coverage.
+
+## Conclusion
+Story 4.1 satisfies non-functional expectations. No blockers identified; continue monitoring CLI smoke results once pipeline extraction begins.
