+++ docs/qa/assessments/4.1-review-20251003.md
@@ -1,0 +1,27 @@
+# QA Review – Story 4.1 (CLI Bridge for FancyRAG kg_build)
+
+Date: 2025-10-03
+Reviewer: Quinn (QA)
+
+## Summary
+- The legacy monolithic CLI has been lifted into `src/fancyrag/cli/kg_build_main.py`, while `scripts/kg_build.py` now serves as a stable wrapper for operators and automation.
+- Parity checks (`artifacts/cli/kg_build_help_before.txt` / `..._after.txt`) and the updated unit suite demonstrate that every flag, default, and semantic toggle survived the refactor.
+- Documentation and risk mitigations were addressed: source tree blueprint reflects the new package and unit tests cover env enforcement plus wrapper delegation.
+
+## Evidence Reviewed
+- Code: `src/fancyrag/cli/kg_build_main.py`, `scripts/kg_build.py`, `docs/architecture/source-tree.md`, story Dev Notes.
+- Tests: `PYTHONPATH=stubs:src pytest tests/unit/fancyrag/cli/test_kg_build_main.py`, `PYTHONPATH=stubs:src pytest tests/unit/scripts/test_kg_build.py`.
+- QA Artifacts: `docs/qa/assessments/4.1-risk-20251003.md`, `docs/qa/assessments/4.1-test-design-20251003.md`, `docs/qa/assessments/4.1-trace-20251003.md`, `docs/qa/assessments/4.1-nfr-20251003.md`.
+
+## Findings
+- Parser defaults, semantic thresholds, and QA flags match the legacy script; unit tests assert `--enable-semantic` and env validation behaviour. ✅
+- Wrapper import path confirmed via unit test; operators continue running `python scripts/kg_build.py` without additional PYTHONPATH tweaks. ✅
+- Documentation updated to announce the packaged CLI module, aligning with Epic 4 guardrails. ✅
+- Residual operational risk (OPS-041) remains low but monitor CI for future module extraction regressions. ⚠️ Monitor
+
+## Recommendation
+- PASS – story ready to merge with standard monitoring of CLI smoke/compose jobs.
+
+## Follow-Ups
+- Automate help-parity snapshot comparison in CI to guard against future parser drift (ties to risk TECH-041).
+
