+++ docs/qa/assessments/4.1-risk-20251003.md
@@ -1,0 +1,68 @@
+# Risk Assessment: Story 4.1 CLI Bridge for FancyRAG kg_build
+
+- **Date:** 2025-10-03
+- **Assessor:** Quinn (Test Architect & Quality Advisor)
+- **Story Reference:** docs/stories/4.1.cli-kg-build-main.md
+- **Epic:** 4 — FancyRAG `kg_build.py` Monolith Decomposition
+- **Overall Story Risk Score:** 70 (Base 100 − 3 × High Risks)
+
+## Executive Summary
+
+- Story refactors the `kg_build` CLI into a packaged entry point while preserving operator behaviour, semantic enrichment toggles, and QA guardrails.
+- Three high risks identified, concentrated in technical and operational domains; no critical (score 9) risks.
+- Mitigations rely on exhaustive parity checks, regression automation, and incremental rollout safeguards (feature flag/branch smoke).
+- Residual risk considered manageable with mandated mitigations; quality gate status = **CONCERNS** pending mitigation execution.
+
+## Risk Matrix
+
+| Risk ID  | Category | Description                                                   | Probability | Impact | Score | Priority |
+|----------|----------|---------------------------------------------------------------|-------------|--------|-------|----------|
+| TECH-041 | Technical | CLI option parity gaps when extracting argparse wiring       | Medium (2)  | High (3) | 6 | High |
+| TECH-042 | Technical | Shared client/env wiring regressions (ensure_env, retries)  | Medium (2)  | High (3) | 6 | High |
+| OPS-041  | Operational | Deployment/runtime breakage from import path/package layout changes | Medium (2) | High (3) | 6 | High |
+
+## Category Distribution
+
+- Technical: 2 high risks (CLI coverage, dependency wiring)
+- Operational: 1 high risk (deployment surface)
+- Security, Performance, Data, Business: No specific risks identified for this story; monitor during integration testing.
+
+## Detailed Risk Register
+
+| Risk ID  | Title | Description | Affected Components | Detection Method | Probability | Impact | Score | Mitigations | Owner |
+|----------|-------|-------------|---------------------|------------------|-------------|--------|-------|-------------|-------|
+| TECH-041 | CLI flag parity regression | Migrating parser/arg defaults into `src/fancyrag/cli/kg_build_main.py` may omit rarely used options (`--include-pattern`, semantic QA thresholds, reset flags) or alter defaults, breaking ingestion workflows. | `scripts/kg_build.py`, new `fancyrag.cli.kg_build_main` | Diff/scan of existing parser, unit snapshot of parsed args, integration smoke | Medium (2) | High (3) | 6 | 1) Generate parser inventory before refactor. 2) Author unit tests asserting every flag/default. 3) Run existing compose smoke to verify runtime parity. | Dev Owner |
+| TECH-042 | Shared client/env wiring regression | Refactor could bypass `ensure_env` loading or `SharedOpenAIClient`, leading to 1) missing `.env` variables at runtime, 2) loss of retry/backoff and sanitization, or 3) secrets logged unintentionally. | `fancyrag.utils.env.ensure_env`, `SharedOpenAIClient`, new CLI module | Unit coverage verifying env resolution, structured logging review, dry-run with missing env | Medium (2) | High (3) | 6 | 1) Centralize env bootstrap inside new module with contract tests (missing var => SystemExit). 2) Add regression test ensuring CLI uses `SharedOpenAIClient`. 3) Lint logs for secret redaction via sanitizer. | Dev Owner |
+| OPS-041 | Import path/entrypoint breakage | Moving logic under `src/fancyrag/` may fail when operators run `python scripts/kg_build.py` due to missing `PYTHONPATH` adjustments, packaging mistakes, or circular imports introduced by new module boundaries. | `scripts/kg_build.py`, packaging config, developer workflows | Local smoke via `python scripts/kg_build.py`, CI pipeline, packaging lints | Medium (2) | High (3) | 6 | 1) Maintain thin wrapper with explicit import + `main()` call plus integration test executed via CLI. 2) Update documentation (source-tree shard) and ensure `PYTHONPATH=src` guidance. 3) Add CI step verifying module import under default env. | Dev Owner |
+
+## Mitigation & Follow-up Plan
+
+1. **Parity Snapshot:** Before refactor, capture current argparse options (`scripts/kg_build.py --help`) and encode into unit tests for new module. Due 2025-10-05.
+2. **Env & Client Regression Tests:** Implement focused unit tests verifying `ensure_env` invocation and `SharedOpenAIClient` usage; include negative tests for missing env to ensure SystemExit. Due 2025-10-06.
+3. **CLI Smoke Extension:** Ensure `tests/integration/local_stack/test_minimal_path_smoke.py` imports the new entry point and runs via script invocation. Due 2025-10-07.
+4. **Documentation Update:** Update `docs/architecture/source-tree.md` and run docs lint guard after module extraction. Due with implementation PR.
+
+## Risk-Based Testing Strategy
+
+- **Priority 1 (High Score 6):**
+  - Unit snapshot tests verifying argument parity (TECH-041).
+  - Unit tests covering `ensure_env`/`SharedOpenAIClient` usage and retry configuration (TECH-042).
+  - CLI invocation integration test (`python scripts/kg_build.py --help`) executed in CI (OPS-041).
+- **Priority 2:** Regression of semantic QA toggles via targeted unit/integration tests ensuring thresholds propagate to evaluator (TECH-042 link).
+- **Priority 3:** Documentation lint & branch smoke verifying operator instructions (OPS-041).
+
+## Residual Risk Assessment
+
+- After mitigations, expected residual score per risk drops to 4 (Medium) due to lower probability (1) with strong automated guards; gate can upgrade to PASS post-evidence review.
+- Document evidence links in QA gate file once mitigations land (unit test file paths, smoke test results).
+
+## Monitoring & Triggers
+
+- Re-run risk profile if additional modules (pipeline, splitters) refactors overlap earlier than planned (Story 4.2+).
+- Monitor CI for CLI smoke/test failures; escalate if parity tests trend red >2 consecutive runs.
+- Validate production run logs after deployment to ensure sanitized output still omits secrets.
+
+---
+
+Risk profile: docs/qa/assessments/4.1-risk-20251003.md
+
