+++ docs/qa/assessments/4.1-test-design-20251003.md
@@ -1,0 +1,78 @@
+# Test Design: Story 4.1 CLI Bridge for FancyRAG kg_build
+
+- **Date:** 2025-10-03
+- **Designer:** Quinn (Test Architect & Quality Advisor)
+- **Story Reference:** docs/stories/4.1.cli-kg-build-main.md
+- **Epic:** 4 — FancyRAG `kg_build.py` Monolith Decomposition
+
+## Test Strategy Overview
+
+- Total test scenarios: 6
+- By level: Unit 3 (50%), Integration 2 (33%), E2E 1 (17%)
+- By priority: P0 = 1, P1 = 5, P2 = 0, P3 = 0
+- Focus areas: CLI argument parity, environment/client wiring, script delegation, semantic QA toggles, end-to-end smoke parity.
+- Risk coverage: Mitigates TECH-041, TECH-042, OPS-041 (see mapping below).
+
+## Test Scenarios by Acceptance Criteria
+
+### AC1 — Package CLI wiring into `src/fancyrag/cli/kg_build_main.py`
+
+| ID | Level | Priority | Test | Justification | Mitigates |
+|----|-------|----------|------|---------------|-----------|
+| 4.1-UNIT-001 | Unit | P1 | Snapshot all argparse options/defaults by instantiating the new parser and comparing against a recorded baseline (including help text for semantic flags). | Ensures no CLI flag regression when migrating wiring. | TECH-041 |
+
+### AC2 — Update `scripts/kg_build.py` to delegate to packaged entry point
+
+| ID | Level | Priority | Test | Justification | Mitigates |
+|----|-------|----------|------|---------------|-----------|
+| 4.1-INT-001 | Integration | P1 | Monkeypatch `fancyrag.cli.kg_build_main.main` and invoke `python -m scripts.kg_build` (via subprocess) to confirm delegation and zero side effects on import. | Validates the thin wrapper executes packaged main and returns exit codes correctly. | OPS-041 |
+| 4.1-E2E-001 | E2E | P1 | Run compose-backed smoke (`tests/integration/local_stack/test_minimal_path_smoke.py`) after refactor to confirm CLI invocation performs full ingestion/retrieval. | Guards against runtime regressions introduced by packaging changes. | TECH-041, OPS-041 |
+
+### AC3 — Preserve environment loading, shared clients, semantic/QA toggles
+
+| ID | Level | Priority | Test | Justification | Mitigates |
+|----|-------|----------|------|---------------|-----------|
+| 4.1-UNIT-002 | Unit | P1 | Validate that parser propagates semantic QA flags (`--enable-semantic`, `--qa-max-*`) into pipeline call signature using dependency injection/mocking. | Ensures toggles continue to reach evaluator/extractor. | TECH-042 |
+| 4.1-UNIT-003 | Unit | P0 | Simulate missing `NEO4J_URI` env and assert CLI raises `SystemExit` with clear message after invoking `ensure_env`. | Prevents silent failures and protects secrets handling. | TECH-042 |
+| 4.1-INT-002 | Integration | P1 | Run `python scripts/kg_build.py --help` in isolated environment with `PYTHONPATH=src` to verify `.env` loading, shared client initialization, and sanitized log output summary (no secrets). | Detects runtime import/env issues and keeps operational contract intact. | TECH-042, OPS-041 |
+
+## Risk Coverage
+
+- **TECH-041:** Covered by 4.1-UNIT-001, 4.1-E2E-001.
+- **TECH-042:** Covered by 4.1-UNIT-002, 4.1-UNIT-003, 4.1-INT-002.
+- **OPS-041:** Covered by 4.1-INT-001, 4.1-INT-002, 4.1-E2E-001.
+
+## Recommended Execution Order
+
+1. `4.1-UNIT-003` (P0) — fail fast on env handling.
+2. `4.1-UNIT-001`, `4.1-UNIT-002` — validate parser parity and flag plumbing.
+3. `4.1-INT-001`, `4.1-INT-002` — confirm delegation and runtime env behaviour.
+4. `4.1-E2E-001` — full smoke to certify operational parity.
+
+## Gate YAML Block
+
+```yaml
+test_design:
+  scenarios_total: 6
+  by_level:
+    unit: 3
+    integration: 2
+    e2e: 1
+  by_priority:
+    p0: 1
+    p1: 5
+    p2: 0
+    p3: 0
+  coverage_gaps: []
+```
+
+## Notes & Follow-ups
+
+- Pair test updates with documentation refresh to keep CLI usage guidance aligned.
+- Leverage parametrized fixtures to reuse parser baseline across unit and integration coverage.
+- Capture CLI `--help` output as artifact for change detection across future stories.
+
+---
+
+Test design matrix: docs/qa/assessments/4.1-test-design-20251003.md
+P0 tests identified: 1
