+++ docs/qa/assessments/4.1-trace-20251003.md
@@ -1,0 +1,24 @@
+# Traceability Assessment – Story 4.1 (CLI Bridge for FancyRAG kg_build)
+
+Date: 2025-10-03
+Assessor: Quinn (QA)
+
+## Summary
+- `src/fancyrag/cli/kg_build_main.py` now owns the full KG build CLI, preserving every flag/default and the semantic enrichment flow from the legacy script.
+- `scripts/kg_build.py` is a thin wrapper that imports the packaged entry point, enabling downstream packaging without altering operator ergonomics.
+- New unit coverage exercises argument defaults, environment enforcement, and wrapper delegation while existing kg_build tests continue to pass against the packaged module.
+
+## Acceptance Criteria Mapping
+
+| AC | Implementation Evidence | Test Coverage | Status |
+|----|-------------------------|---------------|--------|
+| 1  | New module `src/fancyrag/cli/kg_build_main.py` contains the full parser, pipeline wiring, and helper functions migrated from the former script (`src/fancyrag/cli/kg_build_main.py:1-2385`). | `tests/unit/fancyrag/cli/test_kg_build_main.py::test_parse_args_defaults` snapshots parser defaults; existing `tests/unit/scripts/test_kg_build.py` suite exercises the migrated pipeline behaviours against the packaged module. | ✅ |
+| 2  | `scripts/kg_build.py` reduced to wrapper invoking `fancyrag.cli.kg_build_main.main()` (`scripts/kg_build.py:1-9`); source tree documentation updated to list the new module layout (`docs/architecture/source-tree.md:23-74`). | `tests/unit/fancyrag/cli/test_kg_build_main.py::test_scripts_wrapper_exposes_packaged_main` asserts wrapper delegates correctly; recorded `artifacts/cli/kg_build_help_before.txt` and `..._after.txt` demonstrate CLI parity. | ✅ |
+| 3  | Environment validation uses `ensure_env` before pipeline execution and shared OpenAI client wiring preserved (`src/fancyrag/cli/kg_build_main.py:2099-2150`); semantic QA thresholds and toggles carried forward unchanged. | `tests/unit/fancyrag/cli/test_kg_build_main.py::test_parse_args_enable_semantic_sets_flag` and `::test_run_invokes_ensure_env_before_pipeline` verify semantic flag behaviour and env enforcement; full kg_build unit suite continues to validate QA threshold plumbing. | ✅ |
+
+## Additional Notes
+- Help snapshots located in `artifacts/cli/` provide reviewer evidence of argument parity and will support regression checks.
+- Risk mitigations (TECH-041/042, OPS-041) now have corresponding automated coverage per the new unit suite.
+
+## Conclusion
+Traceability COMPLETE. All acceptance criteria have been satisfied with executable evidence and automated validation.
