+++ src/fancyrag/cli/__init__.py
@@ -1,0 +1,5 @@
+"""FancyRAG CLI entrypoints."""
+
+from .kg_build_main import main
+
+__all__ = ["main"]
